\name{FastRCS-package}
\docType{package}
\title{Code to compute the FastRCS regression outlyingness index.}
\description{
Uses the FastRCS algorithm to compute the RCS outlyingness index 
 of regression.
}
\details{
\tabular{ll}{
Package: \tab FastRCS\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2013-01-13\cr
Suggests: \tab mvtnorm\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
FastRCS                 Function to compute the FastRCS regression outlyingness
                       index.
NumStarts                 Internal function used to compute the FastRCS regression
                       outlyingness index.
quanf                     Internal function used to compute the FastRCS regression
                       outlyingness index.
}
}
\author{
Kaveh Vakili [aut, cre],
Maintainer: Kaveh Vakili <kaveh.vakili@wis.kuleuven.be>
}
\keyword{package}
